Разбил на модули js код, подключил через единый файл
